:sparkles: melhoria na condição de destruição do componente

+++ design-system/projects/lib/src/services/command-palette/dialog/dialog.component.ts
@@ -1,4 +1,4 @@
-import {Component, ElementRef, AfterViewInit, ViewChild} from '@angular/core';
+import {Component, ElementRef, AfterViewInit, ViewChild, HostListener, Renderer2} from '@angular/core';
 import {ComponentInjectorService} from '../component-injector.service';
 
 @Component({
@@ -10,18 +10,24 @@ export class DialogComponent implements AfterViewInit {
 
   @ViewChild('rootPanel', {static: true}) rootPanelComponent: ElementRef;
   public itself: any;
+  private unlistener: () => void;
 
   constructor(
-    private componentInjectorService: ComponentInjectorService
+    private componentInjectorService: ComponentInjectorService,
+    private renderer2: Renderer2
   ) {
   }
 
   public ngAfterViewInit(): void {
     this.rootPanelComponent.nativeElement.focus();
-  }
-
-  public onBlur(): void {
-    this.componentInjectorService.removeComponentFromBody(this.itself);
+    this.unlistener = this.renderer2.listen('window', 'focusin', event => {
+      if (!(
+        event.target === this.rootPanelComponent.nativeElement
+        || this.rootPanelComponent.nativeElement.contains(event.target)
+      )) {
+        this.componentInjectorService.removeComponentFromBody(this.itself);
+      }
+    });
   }
 
 }
